resolvendo problemas com nomeclatura e criando tratamento de erro 404

+++ src/controllers/authorController.js
@@ -1,3 +1,4 @@
+import error404 from "../errors/error404.js";
 import { author } from "../models/Author.js";
 
 class authorController {
@@ -19,9 +20,7 @@ class authorController {
       if (authorFound !== null) {
         res.status(200).json(authorFound);
       } else {
-        res.status(404).json({
-          message: " ID não encontrado -  Falha inesperada na requisição",
-        });
+        next(new error404("Falha na requisição: ID do Autor não encontrado"));
       }
     } catch (error) {
       next(error);
@@ -43,8 +42,14 @@ class authorController {
   static async updateAuthor(req, res, next) {
     try {
       const id = req.params.id;
-      await author.findByIdAndUpdate(id, req.body);
-      res.status(200).json({ messege: "Atualizado com sucesso" });
+      const authorFound = await author.findById(id);
+
+      if (authorFound !== null) {
+        await author.findByIdAndUpdate(id, req.body);
+        res.status(200).json({ messege: "Atualizado com sucesso" });
+      } else {
+        new error404("Falha na requisição: ID do Autor não encontrado");
+      }
     } catch (error) {
       next(error);
     }
@@ -53,8 +58,14 @@ class authorController {
   static async deleteAuthor(req, res, next) {
     try {
       const id = req.params.id;
-      await author.findByIdAndDelete(id);
-      res.status(200).json({ messege: "Excluido com sucesso" });
+      const authorFound = await author.findById(id);
+
+      if (authorFound !== null) {
+        await author.findByIdAndDelete(id);
+        res.status(200).json({ messege: "Excluido com sucesso" });
+      } else {
+        new error404("Falha na requisição: ID do Autor não encontrado");
+      }
     } catch (error) {
       next(error);
     }
